wip remodel people entities - entity separation
