+++ Models/Entities/Owners/Owner.cs
@@ -1,0 +1,9 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Models.Entities.Owners
+{
+    public class Owner : EntityModel, IOwner
+    {
+        public Person Person { get; set; } = new Person();
+    }
+}
