--- Models/Entities/People/ITenant.cs
@@ -1,13 +1,0 @@
-namespace real_estate_web_api.Models.Entities.People;
-
-public interface ITenant : IEntityModel
-{
-    string TaxDocument { get; set; }
-    string Address { get; set; }
-    DateTime BirthDate { get; set; }
-    string FirstName { get; set; }
-    string LastName { get; set; }
-    string Mobile { get; set; }
-    double Income { get; set; }
-    bool? InterestedInBuying { get; set; }
-}
