+++ Models/Entities/RealEstates/RealEstate.cs
@@ -1,4 +1,6 @@
+using real_estate_web_api.Models.Entities.Owners;
 using real_estate_web_api.Models.Entities.People;
+using real_estate_web_api.Models.Entities.Realtors;
 using real_estate_web_api.Models.Enumerations;
 
 namespace real_estate_web_api.Models.Entities.RealEstates;
@@ -14,6 +16,6 @@ public class RealEstate : EntityModel, IRealEstate
     public double? SaleAmount { get; set; }
     public bool RentAvailable { get; set; }
     public double? RentAmount { get; set; }
-    public IOwner Owner { get; set; } = new Person();
-    public IRealtor Realtor { get; set; } = new Person();
+    public IOwner Owner { get; set; } = new Owner();
+    public IRealtor Realtor { get; set; } = new Realtor();
 }
