+++ Models/Entities/Realtors/IRealtor.cs
@@ -1,0 +1,8 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Models.Entities.Realtors;
+
+public interface IRealtor : IEntityModel
+{
+    Person Person { get; set; }
+}
