+++ Models/Entities/Rentals/IRental.cs
@@ -1,5 +1,7 @@
 using real_estate_web_api.Models.Entities.RealEstates;
 using real_estate_web_api.Models.Entities.People;
+using real_estate_web_api.Models.Entities.Realtors;
+using real_estate_web_api.Models.Entities.Tenants;
 
 namespace real_estate_web_api.Models.Entities.Rentals;
