+++ Models/Entities/Rentals/Rental.cs
@@ -1,5 +1,7 @@
 using real_estate_web_api.Models.Entities.People;
 using real_estate_web_api.Models.Entities.RealEstates;
+using real_estate_web_api.Models.Entities.Realtors;
+using real_estate_web_api.Models.Entities.Tenants;
 
 namespace real_estate_web_api.Models.Entities.Rentals;
 
@@ -9,6 +11,6 @@ public class Rental : EntityModel, IRental
     public DateTime EndDate { get; set; }
     public double MonthlyAmount { get; set; }
     public IRealEstate RealEstate { get; set; } = new RealEstate();
-    public IRealtor Realtor { get; set; } = new Person();
-    public ITenant Tenant { get; set; } = new Person();
+    public IRealtor Realtor { get; set; } = new Realtor();
+    public ITenant Tenant { get; set; } = new Tenant();
 }
