+++ Models/Entities/Tenants/Tenant.cs
@@ -1,0 +1,11 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Models.Entities.Tenants
+{
+    public class Tenant : EntityModel, ITenant
+    {
+        public Person Person { get; set; } = new Person();
+        public double Income { get; set; }
+        public bool? InterestedInBuying { get; set; }
+    }
+}
